LayoutTests/http/tests/navigation/window-open-cross-origin-then-navigated-back-same-origin.html

<!DOCTYPE html>
<html>
<head>
<script src="/js-test-resources/js-test.js"></script>
</head>
<body>
<script>
description("Tests that process swap on navigation does not break navigating back cross-origin a cross-origin window.");
jsTestIsAsync = true;

if (window.testRunner)
    testRunner.setCanOpenWindows();

onload = function() {
    w = window.open();
    w.opener = null;
    w.location = "http://localhost:8000/navigation/resources/navigate-back-same-origin-helper.html";
    handle = setInterval(() => {
        try {
            if (w.location.href != "about:blank") {
                clearInterval(handle);
                shouldBeEqualToString("w.location.href", "http://127.0.0.1:8000/navigation/resources/otherpage.html");
                finishJSTest();
            }
        } catch(e) {
        }
    }, 10);
}
</script>
</body>
</html>
