LayoutTests/http/tests/navigation/window-open-cross-origin-then-navigated-back-same-origin-expected.txt

Tests that process swap on navigation does not break navigating back cross-origin a cross-origin window.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS w.location.href is "http://127.0.0.1:8000/navigation/resources/otherpage.html"
PASS successfullyParsed is true

TEST COMPLETE
